[refactor] Move Game class to runner.py
